Modified Invoice/Order ID Validation Checks, Added Settings
The checks in place weren't enough to keep users from entering invalid data. A user could enter anything above a certain length and it would validate.
Unfortunately this meant that users could enter things like "Invoice # - 09384029384" and get validated; the bot would be unable to match the hashes after that.

I changed this to check for both min/max length of an order id/invoice #, as well as a strip for "#" and " ". Settings are available for min/max length on both order/invoice.